ENH: removed intel warning 2259

+++ src/triangular_surface_mesh.cpp
@@ -872,7 +872,7 @@ bool besthea::mesh::triangular_surface_mesh::print_ensight_datafiles(
     data_file.write( strings[ 2 ].c_str( ), 80 );
 
     for ( lo j = 0; j < _n_nodes; ++j ) {
-      float value = ( *node_data )[ i ]->get( j );
+      float value = static_cast< float >( ( *node_data )[ i ]->get( j ) );
       data_file.write(
         reinterpret_cast< const char * >( &value ), sizeof( float ) );
     }
@@ -920,7 +920,7 @@ bool besthea::mesh::triangular_surface_mesh::print_ensight_datafiles(
     data_file.write( strings[ 2 ].c_str( ), 80 );
 
     for ( lo j = 0; j < _n_elements; ++j ) {
-      float value = ( *element_data )[ i ]->get( j );
+      float value = static_cast< float >( ( *element_data )[ i ]->get( j ) );
       data_file.write(
         reinterpret_cast< const char * >( &value ), sizeof( float ) );
     }
